[nextest-runner] move execution status types into reporter/events.rs
Part of the external interface.
